Replace __unstableHasSlashCommandReplacements with hasSlashCommandReplacements private selector
