+++ packages/block-editor/src/store/selectors.js
@@ -2153,7 +2153,11 @@ function getInsertUsage( state, id ) {
  *
  * @return {boolean} Whether the given block type is allowed to be shown in the inserter.
  */
-const canIncludeBlockTypeInInserter = ( state, blockType, rootClientId ) => {
+export const canIncludeBlockTypeInInserter = (
+	state,
+	blockType,
+	rootClientId
+) => {
 	if ( ! hasBlockSupport( blockType, 'inserter', true ) ) {
 		return false;
 	}
@@ -2562,49 +2566,6 @@ export const hasInserterItems = ( state, rootClientId = null ) => {
 
 	return hasReusableBlock;
 };
-
-/**
- * Returns whether a block can use the slash command to insert or replace with
- * another block type. Returns false when no other block type can be inserted
- * in the same parent context — for example when the editor's allowedBlockTypes
- * setting or a parent block's allowedBlocks setting restricts insertable types
- * to only the current block, or when the block is inside a content-only section
- * (unsynced pattern / templateLock:'contentOnly') where only content-role blocks
- * are permitted and none are available.
- *
- * @param {Object} state    Editor state.
- * @param {string} clientId Client ID of the block.
- *
- * @return {boolean} Whether the slash command can offer block replacements.
- */
-export const __unstableHasSlashCommandReplacements = createSelector(
-	( state, clientId ) => {
-		const blockName = getBlockName( state, clientId );
-		if ( ! blockName ) {
-			return false;
-		}
-
-		const rootClientId = getBlockRootClientId( state, clientId );
-
-		// Return true as soon as we find any other insertable block type.
-		// canIncludeBlockTypeInInserter already accounts for content-only
-		// section restrictions (#76982) and allowedBlockTypes / parent
-		// allowedBlocks settings (#55378).
-		return getBlockTypes().some(
-			( blockType ) =>
-				blockType.name !== blockName &&
-				canIncludeBlockTypeInInserter( state, blockType, rootClientId )
-		);
-	},
-	( state, clientId ) => {
-		const rootClientId = getBlockRootClientId( state, clientId );
-		return [
-			getBlockName( state, clientId ),
-			getBlockTypes(),
-			...getInsertBlockTypeDependants()( state, rootClientId ),
-		];
-	}
-);
 
 /**
  * Returns the list of allowed inserter blocks for inner blocks children.
