+++ packages/block-editor/src/store/test/selectors.js
@@ -4807,120 +4807,6 @@ describe( 'selectors', () => {
 			expect( canMoveBlock( state, 'child' ) ).toBe( true );
 		} );
 	} );
-
-	describe( '__unstableHasSlashCommandReplacements', () => {
-		const { __unstableHasSlashCommandReplacements } = selectors;
-
-		afterEach( async () => {
-			await dispatch( store ).resetBlocks( [] );
-			await dispatch( store ).updateSettings( {
-				allowedBlockTypes: undefined,
-			} );
-		} );
-
-		it( 'returns true when other block types can be inserted (default case)', async () => {
-			await dispatch( store ).resetBlocks( [
-				{
-					clientId: 'para',
-					name: 'core/test-block-a',
-					attributes: {},
-					innerBlocks: [],
-				},
-			] );
-
-			expect(
-				select( store ).__unstableHasSlashCommandReplacements( 'para' )
-			).toBe( true );
-		} );
-
-		it( 'returns false when allowedBlockTypes restricts to only the current block (#55378)', async () => {
-			await dispatch( store ).updateSettings( {
-				allowedBlockTypes: [ 'core/test-block-a' ],
-			} );
-			await dispatch( store ).resetBlocks( [
-				{
-					clientId: 'para',
-					name: 'core/test-block-a',
-					attributes: {},
-					innerBlocks: [],
-				},
-			] );
-
-			expect(
-				select( store ).__unstableHasSlashCommandReplacements( 'para' )
-			).toBe( false );
-		} );
-
-		it( 'returns false when parent allowedBlocks restricts to only the current block (#55378)', async () => {
-			// Use core/test-block-a as the container — no registered test block
-			// declares it as a required parent, so the allowedBlocks restriction
-			// is the only thing that controls insertion here.
-			await dispatch( store ).resetBlocks( [
-				{
-					clientId: 'container',
-					name: 'core/test-block-a',
-					attributes: {},
-					innerBlocks: [
-						{
-							clientId: 'para',
-							name: 'core/test-block-b',
-							attributes: {},
-							innerBlocks: [],
-						},
-					],
-				},
-			] );
-			await dispatch( store ).updateBlockListSettings( 'container', {
-				allowedBlocks: [ 'core/test-block-b' ],
-			} );
-
-			expect(
-				select( store ).__unstableHasSlashCommandReplacements( 'para' )
-			).toBe( false );
-		} );
-
-		it( 'returns false when paragraph is inside a content-only section (unsynced pattern) (#76982)', () => {
-			// A block with templateLock:'contentOnly' whose parent does NOT
-			// also have contentOnly is treated as a section block. Inside it,
-			// canIncludeBlockTypeInInserter returns false for any block type
-			// that isn't a content-role block — so the slash inserter would
-			// show nothing useful and the placeholder should not hint at it.
-			const state = {
-				blocks: {
-					byClientId: new Map( [
-						[ 'section', { name: 'core/test-block-b' } ],
-						[ 'para', { name: 'core/test-block-a' } ],
-					] ),
-					attributes: new Map( [
-						[ 'section', {} ],
-						[ 'para', {} ],
-					] ),
-					parents: new Map( [
-						[ 'section', '' ],
-						[ 'para', 'section' ],
-					] ),
-					order: new Map( [
-						[ '', [ 'section' ] ],
-						[ 'section', [ 'para' ] ],
-					] ),
-					blockEditingModes: new Map(),
-				},
-				blockListSettings: {
-					section: { templateLock: 'contentOnly' },
-				},
-				settings: {
-					[ sectionRootClientIdKey ]: '',
-				},
-				derivedBlockEditingModes: new Map(),
-				preferences: { insertUsage: {} },
-				editedContentOnlySection: null,
-			};
-
-			expect(
-				__unstableHasSlashCommandReplacements( state, 'para' )
-			).toBe( false );
-		} );
-	} );
 } );
 
 describe( 'getInserterItems with core blocks prioritization', () => {
